Issue number: Obtained from: Submitted by:  mranga Reviewed by:   mranga Incorporating the latest API changes from Phelim CVS: ---------------------------------------------------------------------- CVS: Issue number: CVS:   If this change addresses one or more issues, CVS:   then enter the issue number(s) here. CVS: Obtained from: CVS:   If this change has been taken from another system, CVS:   then name the system in this line, otherwise delete it. CVS: Submitted by: CVS:   If this code has been contributed to the project by someone else; i.e., CVS:   they sent us a patch or a set of diffs, then include their name/email CVS:   address here. If this is your work then delete this line. CVS: Reviewed by: CVS:   If we are doing pre-commit code reviews and someone else has CVS:   reviewed your changes, include their name(s) here. CVS:   If you have not had it reviewed then delete this line.

+++ src/examples/authorization/ShootistAuth.java
@@ -28,7 +28,7 @@ public class ShootistAuth implements SipListener {
 	private ListeningPoint udpListeningPoint;
 	private ClientTransaction inviteTid;
 	private Dialog dialog;
-	int invco = 1;
+	long invco = 1;
         String peerHostPort = "127.0.0.1:5070";
         String transport = "udp";
         String USER_AUTH = "auth";
